adjustments to new Fitting folder
