+++ examples_old/example3.py
@@ -5,7 +5,7 @@ from typing import Callable
 
 import numpy as np
 
-from easyscience.Fitting.Fitting import Fitter
+from easyscience.Fitting import Fitter
 from easyscience.Objects.ObjectClasses import BaseObj
 from easyscience.Objects.ObjectClasses import Parameter
