+++ src/easyscience/Fitting/__init__.py
@@ -1,1 +1,4 @@
+from .fitter import Fitter  # noqa: F401, E402
 from .minimizers.fitting_template import FitResults  # noqa: F401, E402
+# Causes a circular import
+# from .multi_fitter import MultiFitter  # noqa: F401, E402
