+++ tests/integration_tests/test_undoRedo.py
@@ -265,7 +265,7 @@ def test_fittingUndoRedo(fit_engine):
 
     y = l1(x) + 0.125 * (dy - 0.5)
 
-    from easyscience.Fitting.Fitting import Fitter
+    from easyscience.Fitting import Fitter
 
     f = Fitter(l2, l2)
     try:
